+++ CHANGELOG.md
@@ -6,6 +6,15 @@
 
 ### Fixed
 
+## [1.12.1]
+
+### Changed
+- general refactor of documentation
+- merge docker compose files into one
+
+### Fixed
+- dns resolution for ipv6
+
 ## [1.12.0]
 
 ### Changed